+++ src/api/services/Login.services.ts
@@ -1,0 +1,10 @@
+import { LoginParams } from "../../types/Login.types";
+import { instance } from "../instance";
+
+export const LoginService = async (user: LoginParams) => {
+  const res = await instance.post("/member/signin", {
+    email: user.email,
+    password: user.password,
+  });
+  return res.data;
+};
