+++ src/pages/Login.tsx
@@ -3,13 +3,13 @@ import { CommonInput, CommonButton } from "../styles/globalStyles";
 import { useNavigate } from "react-router";
 import { LoginSignUpContainer } from "../styles/LoginSignUp.style";
 import { useState, ChangeEvent } from "react";
-import { LoginService } from "../api/services/Main.services";
+import { LoginService } from "../api/services/Login.services";
 
 export default function Login() {
+  const navigate = useNavigate();
+
   const [email, setEmail] = useState<string>("abcd");
   const [password, setPassword] = useState<string>("dddd");
-
-  const navigate = useNavigate();
 
   const goToSignUp = () => {
     navigate("/signUp");
@@ -29,8 +29,18 @@ export default function Login() {
     }
   };
 
-  const handleLogin = async () => {
-    const res = await LoginService({ email, password });
+  const handleLogin = () => {
+    (async () => {
+      try {
+        const res = await LoginService({ email, password });
+        const { accessToken, refreshToken } = res;
+        localStorage.setItem("access", accessToken);
+        localStorage.setItem("refresh", refreshToken);
+        navigate("/main");
+      } catch (error) {
+        console.error("로그인 실패:", error);
+      }
+    })();
   };
 
   return (
